Move Node, NodeDriver and associated classes from feditest.protocols to feditest.nodedrivers -- much more logical location
