+++ tests.unit/test_20_create_test_plan_session_template.py
@@ -7,7 +7,7 @@ import pytest
 
 import feditest
 from feditest import test
-from feditest.protocols import Node
+from feditest.nodedrivers import Node
 from feditest.testplan import TestPlanConstellation, TestPlanConstellationNode, TestPlanSession, TestPlanTestSpec
